+++ SomnusBackend/src/main/java/com/somnus/server/backend/articles/domain/ArticleTopic.java
@@ -1,5 +1,15 @@
 package com.somnus.server.backend.articles.domain;
 
 public enum ArticleTopic {
-    MATHEMATICS, PHYSICS, ENGINEERING, CHEMISTRY
+    MATHEMATICS("Mathematics"),
+    PHYSICS("Physics"),
+    ENGINEERING("Engineering"),
+    CHEMISTRY("Chemistry"),
+    GAMING("Gaming");
+
+    public final String name;
+
+    ArticleTopic(String name) {
+        this.name = name;
+    }
 }